Modify: types, stores 폴더 내 파일 분리 및 import 적용
