+++ src/components/analyze/CodeViewer.tsx
@@ -1,11 +1,13 @@
 "use client";
 
-import { useFileViewerStore } from "@/stores/store";
+import { useFileViewerStore } from "@/stores/useFileViewerStore";
 
-import { IconMagnifierWithPlus } from "../ui/Icons";
 import { getLanguage } from "@/lib/utils";
 import dynamic from "next/dynamic";
 import { useEffect, useState } from "react";
+import { Alert } from "../ui/Alert";
+import { IconMagnifierWithPlus } from "../ui/Icons";
+import { useFileProcessStore } from "@/stores/useFileProcessStore";
 
 const SyntaxHighlighter = dynamic(
   () => import("react-syntax-highlighter").then((mod) => mod.Prism),
@@ -14,13 +16,25 @@ const SyntaxHighlighter = dynamic(
 
 export default function CodeViewer() {
   const { currentFile, fileContent, isLoading, error } = useFileViewerStore();
+  const fileStatuses = useFileProcessStore((state) => state.fileStatuses);
+  const status = Array.from(fileStatuses.values())[0];
+
   const [highlighterStyle, setHighlighterStyle] = useState({});
+  const [hasAlertBeenSet, setHasAlertBeenSet] = useState(false);
+  const [isAlertOpen, setIsAlertOpen] = useState(false);
 
   useEffect(() => {
     import("react-syntax-highlighter/dist/esm/styles/prism/one-light").then(
       (mod) => setHighlighterStyle(mod.default),
     );
   }, []);
+
+  useEffect(() => {
+    if (status && !hasAlertBeenSet) {
+      setIsAlertOpen(true);
+      setHasAlertBeenSet(true);
+    }
+  }, [status, hasAlertBeenSet]);
 
   const renderContent = (): string => {
     if (isLoading) {
@@ -48,7 +62,7 @@ export default function CodeViewer() {
   }
 
   return (
-    <div className="w-full overflow-hidden rounded-lg border border-[#c3c3c3]">
+    <div className="relative w-full overflow-hidden rounded-lg border border-[#c3c3c3]">
       {/* <ProcessStatus status={status} /> */}
       <SyntaxHighlighter
         language={currentFile ? getLanguage(currentFile) : "text"}
@@ -64,6 +78,7 @@ export default function CodeViewer() {
       >
         {renderContent()}
       </SyntaxHighlighter>
+      {isAlertOpen && <Alert status={status} />}
     </div>
   );
 }
