+++ src/components/analyze/FileExplorer.tsx
@@ -1,10 +1,11 @@
 "use client";
 
+import { expandFolder } from "@/lib/api/repositories";
+import { useFileSelectionStore } from "@/stores/useFileSelectionStore";
+import { useFileViewerStore } from "@/stores/useFileViewerStore";
+import { RepoContentItem } from "@/types/repo";
 import { useCallback, useEffect, useState } from "react";
 import FileList from "./FileList";
-import { RepoContentItem } from "@/types/type";
-import { expandFolder } from "@/lib/api/repositories";
-import { useFileViewerStore, useFileSelectionStore } from "@/stores/store";
 
 export default function FileExplorer({
   initialStructure,
