+++ src/components/vulnerability-db/ArticleDetail.tsx
@@ -1,4 +1,4 @@
-import { ArticleDetailProps } from "@/types/type";
+import { ArticleDetailProps } from "@/types/post";
 import { IconPin, IconShare } from "../ui/Icons";
 import { Label } from "../ui/Label";
