+++ src/components/vulnerability-db/VulDBDashboard.tsx
@@ -14,7 +14,7 @@ import { Ranking } from "@/components/ui/Ranking";
 import { loginWithGithub } from "@/lib/actions";
 import { ITEMS_PER_DB_PAGE, PAGES_PER_GROUP } from "@/lib/const";
 import { formatTimestampAsDaysAgo } from "@/lib/utils";
-import { VulDBPost } from "@/types/type";
+import { VulDBPost } from "@/types/post";
 import { isCertCCContentType } from "@/types/typeGuards";
 import { SessionProvider, useSession } from "next-auth/react";
 import { useState, useTransition } from "react";
