+++ src/components/vulnerability-db/VulDBImageCardContainer.tsx
@@ -1,5 +1,5 @@
 import { cn, formatTimestampAsDateTime } from "@/lib/utils";
-import { VulDBPost } from "@/types/type";
+import { VulDBPost } from "@/types/post";
 import Link from "next/link";
 import {
   Card,
