+++ src/types/file.d.ts
@@ -1,0 +1,1 @@
+export type FileStatus = "onCheck" | "onWait" | "error" | "success" | null;
